[Fix][API] Fix column length can not be long (#8039)

+++ seatunnel-api/src/main/java/org/apache/seatunnel/api/table/catalog/schema/ReadonlyConfigParser.java
@@ -131,14 +131,12 @@ public class ReadonlyConfigParser implements TableSchemaParser<ReadonlyConfig> {
                                                                 new IllegalArgumentException(
                                                                         "schema.columns.* config need option [type], please correct your config first"));
 
-                                Integer columnLength =
+                                Long columnLength =
                                         columnConfig.get(
                                                 TableSchemaOptions.ColumnOptions.COLUMN_LENGTH);
-
                                 Integer columnScale =
                                         columnConfig.get(
                                                 TableSchemaOptions.ColumnOptions.COLUMN_SCALE);
-
                                 Boolean nullable =
                                         columnConfig.get(TableSchemaOptions.ColumnOptions.NULLABLE);
                                 Object defaultValue =
@@ -149,7 +147,7 @@ public class ReadonlyConfigParser implements TableSchemaParser<ReadonlyConfig> {
                                 return PhysicalColumn.of(
                                         name,
                                         seaTunnelDataType,
-                                        Long.valueOf(columnLength),
+                                        columnLength,
                                         columnScale,
                                         nullable,
                                         defaultValue,
